ampro-python/pyproject.toml
[tool.poetry]
name = "ampro-python"
version = "0.1.0"
description = ""
authors = ["Stubbs <stubbs0x@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
